+++ pages/assignment-creator.tsx
@@ -7,11 +7,12 @@ import AssignmentCreationForm, {
 } from "../components/assignment-creator/assignmentCreationForm";
 import Navbar from "../components/Navbar";
 import { Question } from "./api/question";
-import { useMutation } from "@apollo/client";
+import { ApolloClient, InMemoryCache, useMutation } from "@apollo/client";
 import AssignmentConfirmation from "../components/assignment-creator/assignmentConfirmation";
 import { CREATE_ASSIGNMENT } from "../graphql/createAssignment";
 import DisplayAssignmentQuestions from "../components/assignment-creator/displayAssignmentQuestions";
 import { QuestionType } from "./api/questionTypes";
+import { FETCH_SKILL_DESCRIPTION_AND_GRADE } from "../graphql/fetchSkillDescriptionAndGrade";
 
 enum STAGE {
   CHOOSE_SKILLS,
@@ -19,7 +20,7 @@ enum STAGE {
   CONFIRM,
 }
 
-const Diagnostic = () => {
+const AssignmentCreator = (data) => {
   const [stage, setStage] = useState(STAGE.CHOOSE_SKILLS);
   const [skills, setSkills] = useState<Skill[]>([]);
   const [questionTypes, setQuestionTypes] = useState<QuestionType[]>([]);
@@ -75,6 +76,7 @@ const Diagnostic = () => {
     case STAGE.CHOOSE_SKILLS:
       component = (
         <AssignmentCreationForm
+          data={data}
           onClick={createAssignment}
           questionCounts={questionCounts}
           setQuestionCounts={setQuestionCounts}
@@ -109,4 +111,28 @@ const Diagnostic = () => {
   );
 };
 
-export default Diagnostic;
+export async function getServerSideProps() {
+  const client = new ApolloClient({
+    uri: "https://talented-duckling-40.hasura.app/v1/graphql/",
+    cache: new InMemoryCache(),
+  });
+
+  const { data } = await client.query({
+    query: FETCH_SKILL_DESCRIPTION_AND_GRADE,
+  });
+
+  if (!data) {
+    return {
+      redirect: {
+        destination: "/",
+        permanent: false,
+      },
+    };
+  }
+
+  return {
+    props: data, // will be passed to the page component as props
+  };
+}
+
+export default AssignmentCreator;
